+++ examples/water_boiler/water_boiler.py
@@ -1,8 +1,6 @@
-#!/usr/bin/env python
-
-import time
-import matplotlib.pyplot as plt
-from simple_pid import PID
+import utime
+from control import PID
+from machine import Timer
 
 
 class WaterBoiler:
@@ -17,45 +15,34 @@ class WaterBoiler:
     def update(self, boiler_power, dt):
         if boiler_power > 0:
             # Boiler can only produce heat, not cold
-            self.water_temp += 1 * boiler_power * dt
+        self.water_temp += 1 * boiler_power * dt / 1000  # just to convert from miliseconds to seconds
 
         # Some heat dissipation
-        self.water_temp -= 0.02 * dt
+        self.water_temp -= 0.2 * dt
         return self.water_temp
 
+interval1 = 100
+boiler = WaterBoiler()
+water_temp = boiler.water_temp
+pid = PID(1, 0, 0, setpoint=20, sample_time=interval1, scale='ms')
+pid.output_limits = (0, 100)
 
-if __name__ == '__main__':
-    boiler = WaterBoiler()
-    water_temp = boiler.water_temp
+#Timer Function Callback
+def timerFunc0(t):
+    global power
+    global water_temp
+    power = pid(water_temp)
+    water_temp = boiler.update(power, interval1)
+    print(pid._last_time,",",water_temp,",",pid.setpoint,",",power)
 
-    pid = PID(5, 0.01, 0.1, setpoint=water_temp)
-    pid.output_limits = (0, 100)
+#Timer Function Callback
+def timerFunc1(t):
+    global pid
+    pid.setpoint = 100
 
-    start_time = time.time()
-    last_time = start_time
 
-    # Keep track of values for plotting
-    setpoint, y, x = [], [], []
+tim0=Timer(0)
+tim0.init(period=100, mode=Timer.PERIODIC, callback=timerFunc0)
+tim1=Timer(1)
+tim1.init(period=5000, mode=Timer.PERIODIC, callback=timerFunc1)
 
-    while time.time() - start_time < 10:
-        current_time = time.time()
-        dt = current_time - last_time
-
-        power = pid(water_temp)
-        water_temp = boiler.update(power, dt)
-
-        x += [current_time - start_time]
-        y += [water_temp]
-        setpoint += [pid.setpoint]
-
-        if current_time - start_time > 1:
-            pid.setpoint = 100
-
-        last_time = current_time
-
-    plt.plot(x, y, label='measured')
-    plt.plot(x, setpoint, label='target')
-    plt.xlabel('time')
-    plt.ylabel('temperature')
-    plt.legend()
-    plt.show()