+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [1.1.0] - 2021-07-17
+
+### Changes
+
+- Forked from [m-lundberg/simple-pid](https://github.com/m-lundberg/simple-pid)
+
+- Ported from Python to MicroPython
+
+
 ## [1.0.1] - 2021-04-11
 
 ### Fixed
@@ -89,7 +98,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation
 
-[Unreleased]: https://github.com/m-lundberg/simple-pid/compare/v1.0.1...HEAD
+[Unreleased]: https://github.com/JorgeGMarques/simple-pid/compare/v1.1.1...HEAD
+[1.1.0]: https://github.com/JorgeGMarques/simple-pid/compare/v1.0.1...v1.1.0
 [1.0.1]: https://github.com/m-lundberg/simple-pid/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/m-lundberg/simple-pid/compare/v0.2.4...v1.0.0
 [0.2.4]: https://github.com/m-lundberg/simple-pid/compare/v0.2.3...v0.2.4